Use average affinity when building clusters
aff_local is the running total of affinity from
each site to the sites in the cluster.

The previous approach was to take the mean of running affinity
totals within the cluster, scale it by aff_thres, and compare
with the running affinity totals of sites outside the cluster.
The site was included if it's running affinity was growing faster than
the average running affinity within the cluster, with a handicap off aff_thres
for sites already in the cluster.

Ben-Dor et al. 1999 seem to have just taken the
average affinity of a new site to the current sites in
the cluster, and checked if it was greater than threshold.
Now a site is included if it is sufficiently associated
with the cluster, ie. mean affinity to cluster >= aff_thres.