Move dens and X data out of individual bootstrapped GFs and CGFs
dens and X are identical between each bootstrap run, but are stored
by every gradientforest and combinedgradientforest object within the
bootstrap. They are now kept once as part of the bootstrapped
gradientforest of bootstrapped combinedgradientforest.

Care does need to be taken when plotting with gradientforest
functions, because the gradientforest package sometimes expects dens
and X to be present in the gradientforest object.