Fix bug with variable name collision. Re-add parallel processing
X is a parameter to lapply, so passing X in as an extra arg to FUN
breaks lapply.

Now that gf objects have X and dens stripped out, the setup time for
parallel workers is much lower, and working in parallel speeds up
creation of combined bootstrapped gradient forest objects.